Ft/service pack module

+++ FrameMates_API/src/main/java/fu/training/FrameMates_API/studio/StudioModel.java
@@ -1,5 +1,7 @@
 package fu.training.FrameMates_API.studio;
 
+import fu.training.FrameMates_API.servicepack.ServicePack;
+import fu.training.FrameMates_API.servicepack.ServicePackModel;
 import lombok.Data;
 
 import java.sql.Timestamp;
@@ -10,9 +12,10 @@ public class StudioModel {
     private String address;
     private String description;
     private Integer status;
-    private Integer totalRating;
+    private Double totalRating;
     private Integer balance;
     private Timestamp createDate;
     private String avatarStudio;
     private String coverImage;
+
 }
